Simplix_LS2
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3668 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: f4b6a6a9b22452204d2933ffd1c6836fabc9dc95
Former-commit-id: fef07d1a7a68c59d55f3212e87b21e645889249c

+++ src/drivers/simplix/src/unitclothoid.cpp
@@ -366,7 +366,7 @@ void TClothoidLane::SmoothBetween(int Step, double BumpMod)
 	  TVec3d P3 = L3->Point;
 
 	  double T = L0->Offset + L1->Offset + L2->Offset;
-	  L1->Offset = T/3;
+	  L1->Offset = (float) (T/3);
     }
   }
 }
@@ -427,9 +427,9 @@ void TClothoidLane::SetOffset
 
   if (!(P->Fix))
   {
-    P->Offset = T;
+    P->Offset = (float) T;
     P->Point = P->CalcPt();
-    P->Crv = TUtils::CalcCurvatureXY(PP->Point, P->Point, PN->Point);
+    P->Crv = (float) TUtils::CalcCurvatureXY(PP->Point, P->Point, PN->Point);
   }
 }
 //==========================================================================*
